Moved appmodule pageloads into their own folder
